Add 32bit support

+++ XR_APILAYER_NOVENDOR_motion_compensation/input.cpp
@@ -483,7 +483,7 @@ namespace input
                                TLArg(index, "Index"));
 
         std::string path;
-        if (const auto buttons = m_Mapping.find(profile); buttons != m_Mapping.end() && buttons->second.size() > index)
+        if (const auto buttons = m_Mapping.find(profile); buttons != m_Mapping.end() && (int) buttons->second.size() > index)
         {
             path = buttons->second.at(index);
         }
